50 fib numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -12,16 +12,16 @@
 
 int main(void)
 {
-	unsigned long long int;
-	unsigned long long int fib[50] = {1, 2};
+	unsigned long int;
+	unsigned long int fib[50] = {1, 2};
 
-	printf("%llu, %llu", fib[0], fib[1]);
+	printf("%lu, %lu", fib[0], fib[1]);
 
 	i = 2;
 	while (i < 50)
 	{
 		fib[i] = fib[i - 1] + fib[i - 2];
-		printf(" ,%llu", fib[i]);
+		printf(" ,%lu", fib[i]);
 		i++;
 	}
 	printf("\n");
